chore(snapshot): full current build of push-snapshot
Squashed snapshot of the working branch so the remote carries the complete
current state as one clean commit on top of origin/main.

Highlights (relative to origin/main):
- manipulation delivery: per-TF target caps (1d/4h/1h), 1w macro ladder, TP1-RR
  gate, confirmed-short gate (HUNT_MANIP_SHORT_REQUIRE_LTF), laddered-TP tracking,
  ladder доборы + partial-fix campaign model (prod + research/backtest_scanner MAE).
- prizrak: market-cap доп-фактor (marketcap*.py), corpus + razbor tooling.
- diagnostics/observability: universe_health, runtime/telemetry, docs/ARCHITECTURE.
- tooling: .claude/.opencode skills, .mcp.json, .cursor rules.

Security: .env.example scrubbed to PLACEHOLDERS (real tokens removed; they live only
in gitignored .env). No secret is present in this commit's tree or in any pushed
history — verified by a staged-content secret gate before commit.

+++ hunt_core/track/path_backfill.py
@@ -56,13 +56,13 @@ def compute_derived_from_path(
         }
 
     best_favorable = 0.0
-    best_favorable_ts: int | None = None
+    best_favorable_ts: float | None = None
     worst_adverse = 0.0
-    worst_adverse_ts: int | None = None
+    worst_adverse_ts: float | None = None
     # First-touch tracking: for each ATR threshold, the first bar whose
     # high/low crosses it (adverse resolved first within a bar — loss-first
     # tie-break, see module docstring).
-    first_touch: dict[float, int | None] = {g: None for g in _FIRST_PASSAGE_ATR_GRID}
+    first_touch: dict[float, float | None] = {g: None for g in _FIRST_PASSAGE_ATR_GRID}
     ret_at_offsets: dict[str, float] = {}
     offsets_remaining = {m: (m, f"+{m}m" if m < 60 else f"+{m // 60}h") for m in _RET_OFFSETS_MIN}
